requirements.txt
discord.py>=2.3.0
Pillow>=10.0.0
aiohttp>=3.8.0
colorama>=0.4.6
python-dotenv>=1.0.0